+++ src/main/java/de/skuzzle/test/snapshots/data/SnapshotSerializer.java
@@ -1,6 +1,0 @@
-package de.skuzzle.test.snapshots.data;
-
-public interface SnapshotSerializer {
-
-    String serialize(Object testResult) throws SnapshotException;
-}