+++ src/main/java/de/skuzzle/test/snapshots/data/StructuralAssertions.java
@@ -1,6 +1,0 @@
-package de.skuzzle.test.snapshots.data;
-
-public interface StructuralAssertions {
-
-    void assertEquals(String storedSnapshot, String serializedActual) throws AssertionError, SnapshotException;
-}