+++ src/main/java/de/skuzzle/test/snapshots/data/xml/XmlUnitStructuralAssertions.java
@@ -2,9 +2,9 @@ package de.skuzzle.test.snapshots.data.xml;
 
 import org.xmlunit.assertj.XmlAssert;
 
-import de.skuzzle.test.snapshots.data.StructuralAssertions;
+import de.skuzzle.test.snapshots.StructuralAssertions;
 
-public class XmlUnitStructuralAssertions implements StructuralAssertions {
+class XmlUnitStructuralAssertions implements StructuralAssertions {
 
     @Override
     public void assertEquals(String storedSnapshot, String serializedActual) {