+++ RELEASE_NOTES.md
@@ -1,4 +1,7 @@
-* Initial
+* [#2](https://github.com/skuzzle/snapshot-tests/issues/2): Allow to access some snapshot information from within the test case.
+* [#4](https://github.com/skuzzle/snapshot-tests/issues/4): Retain original stack trace on assertion failure
+* Internal refactoring
+* Don't rely on spring-boot dependency management anymore
 
 Maven Central coordinates for this release:
 
@@ -6,7 +9,7 @@ Maven Central coordinates for this release:
 <dependency>
     <groupId>de.skuzzle.test</groupId>
     <artifactId>snapshot-tests</artifactId>
-    <version>0.0.1</version>
+    <version>0.0.2</version>
     <scope>test</scope>
 </dependency>
 ```